@@ -147,7 +147,7 @@
     "react-native": "local-cli/wrong-react-native.js"
   },
   "peerDependencies": {
-    "react": "16.5.0"
+    "react": "16.6.0-alpha.400d197"
   },
   "dependencies": {
     "@babel/runtime": "^7.0.0",
@@ -223,9 +223,9 @@
     "jest": "23.4.1",
     "jest-junit": "5.1.0",
     "prettier": "1.13.6",
-    "react": "16.5.0",
+    "react": "16.6.0-alpha.400d197",
     "react-native-dummy": "0.1.0",
-    "react-test-renderer": "16.5.0",
+    "react-test-renderer": "16.6.0-alpha.400d197",
     "shelljs": "^0.7.8",
     "sinon": "^2.2.0"
   },
